requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0

# Async support
aiohttp>=3.8.0
asyncio>=3.4.3

# PDF generation
reportlab>=4.0.5
Pillow>=10.0.0

# AI Models
google-generativeai>=0.3.2
anthropic>=0.23.1

# Database
supabase>=2.5.5
psycopg2-binary>=2.9.9

# AWS S3
boto3>=1.34.11
aioboto3>=12.3.0

# FastAPI dependencies
fastapi==0.100.0
hypercorn==0.14.4
pydantic>=2.0.2
python-multipart>=0.0.6
